Issue #849
@bsaiken @gsouvik

Bug fix: in the mobile map, the collect data only worked when selecting
a task location in the list. Selecting a task location on the map didn't
store the task location id to use it to get the forms associated.
Now the task location id is stored in a hidden field so we can from both
map or list have the same functions.